ENH: Finished backend for #41; docstrings still needed

+++ src/chemcoord/internal_coordinates/_zmat_class_core.py
@@ -642,7 +642,7 @@ and assigning values safely.
     def get_grad_X(self, as_function=True):
         zmat = self.change_numbering()
         c_table = zmat.loc[:, ['b', 'a', 'd']].values.T
-        C = zmat.loc[:, ['bond', 'angle', 'dihedral']].values.T
+        C = zmat.loc[:, ['bond', 'angle', 'dihedral']].values.astype('f8').T
         C[[1, 2], :] = np.radians(C[[1, 2], :])
 
         grad_X = transformation.get_grad_X(C, c_table)
